Use `is-absolute-url` package to decide external links
Currently `.`, `..` and `example.md` are treated as external links.  It
looks very odd.

This package decides external links by the porr regexp. It is the reason
of the above problem. Fixing regexp for this is good idea, but, I think
regexp mantenance is hard costed, so using expert library is also good idea.
`is-absolute-url` is popular library for this purpose. It is downloaded 1M+
times in a week.